+++ sw/device/tests/sram_ctrl_sleep_sram_ret_contents_impl.c
@@ -155,9 +155,10 @@ void set_up_reset_request(void) {
 
   CHECK_DIF_OK(dif_aon_timer_wakeup_stop(&aon_timer));
 
-  // Enter low power mode.
-  CHECK_STATUS_OK(
-      aon_timer_testutils_watchdog_config(&aon_timer, UINT32_MAX, 20, false));
+  // Enter low power mode. Use UINT32_MAX as wakeup threshold as UINT64_MAX far
+  // too long a timeout.
+  CHECK_STATUS_OK(aon_timer_testutils_watchdog_config(
+      &aon_timer, (uint64_t)UINT32_MAX, 20, false));
   LOG_INFO("wait for reset");
   wait_for_interrupt();
   CHECK(false, "Should have a reset to CPU and ret_sram before this line");
